+++ CHANGELOG.md
@@ -15,6 +15,9 @@ This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 - Don't hide dropdown on blur event by default.
 - Don't activate the first dropdown item by default.
 
+### Removed
+- Remove `Dropdown#length`
+
 ## [0.2.0] - 2016-02-29
 ### Added
 - Enable to select dropdown in touch devices.